Redesigned the coed file structure.
Once upon a time..

There was an ICoedObject interface that contains all
the versioning and collabortion methods inherited from
IVersionedObject and ICollabObject.
These objects were returned/created by the ICoedCommunicator's
getObject and addObject methods which took a path relative
to the workspace as a parameter (either a file or a folder),
the difference being that if getObject is called for an
unversioned file, its versioning component is left "null",
while addObject adds the file to versioning.
If getObject is called for a folder then all actions
are done recursively on the files in that folder. In the case
of versioning methods, they only apply to versioned files
while collaboration methods apply to all files.

The END.